Restructure tests back to per-topic files
The previous sweep collapsed every topic into a single test/lib/shared.ts,
losing the per-file granularity that the original master/test/*.ts layout
offered. This commit decomposes shared.ts into topic-specific modules under
test/lib/ (tee, method, chunked, status, sub-path), each exporting a single
run<Topic>Tests(label, express) function. Common helpers (cache prefix
factory and the before/after cleanup hooks) live in test/lib/util.ts.

The two entry points (express-tee.express4.test.ts /
express-tee.express5.test.ts) import all five run<Topic>Tests and invoke
them in sequence, so the Express 4/5 alias-driven parallelism is preserved
while the topic granularity from master is restored.
